Laser Image Analysis in ImageJ
A set of plugins/macros I developed for particle analysis of images of
laser points and bright spots. By measuring distance between laser
points and bright spots, we were able to probe at the microstructure of
the solar concentrator prototypes. ImageJ is an image analysis
application written in Java. The .ijm files are written in the ImageJ
macro language.